refactor: Cash 필드 long으로 형변환 처리

+++ src/main/java/lotto/domain/LotteryShop.java
@@ -4,7 +4,7 @@ import lotto.domain.lotto.AutoLottoGenerator;
 import lotto.domain.lotto.Lottos;
 
 public class LotteryShop {
-    public static final int LOTTO_PRICE = 1000;
+    public static final long LOTTO_PRICE = 1000L;
 
     private LotteryShop() {}
