+++ src/main/java/lotto/domain/lotto/AutoLottoGenerator.java
@@ -1,14 +1,11 @@
 package lotto.domain.lotto;
 
-import lotto.domain.lotto.Lotto;
-import lotto.domain.lotto.LottoNumber;
-import lotto.domain.lotto.Lottos;
-
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
 import java.util.stream.Collectors;
 import java.util.stream.IntStream;
+import java.util.stream.LongStream;
 
 public class AutoLottoGenerator {
     private static final List<LottoNumber> BASE_NUMBERS;
@@ -21,8 +18,8 @@ public class AutoLottoGenerator {
 
     private AutoLottoGenerator() {}
 
-    public static Lottos tickets(int amount) {
-        final List<Lotto> lottos = IntStream.range(0, amount)
+    public static Lottos tickets(long amount) {
+        final List<Lotto> lottos = LongStream.range(0, amount)
                 .mapToObj(idx -> ticket())
                 .collect(Collectors.toList());
